sql_to_postgres/sqlite_to_postgres/requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.0.3
